+++ src/main/java/roomescape/config/AuthInfoArgumentResolver.java
@@ -2,13 +2,15 @@ package roomescape.config;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
+import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.service.AuthService;
+import roomescape.service.auth.AuthService;
 import roomescape.service.dto.AuthInfo;
 
+@Component
 public class AuthInfoArgumentResolver implements HandlerMethodArgumentResolver {
 
     private final AuthService authService;
@@ -26,11 +28,6 @@ public class AuthInfoArgumentResolver implements HandlerMethodArgumentResolver {
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
                                   NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
         HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
-
-        if (request.getCookies() == null) {
-            return null;
-        }
-
         return authService.getAuthInfo(request.getCookies());
     }
 }
